requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
Django==2.2a1
djangorestframework==3.9.1
idna==2.7
jsonschema==3.0.1
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==4.3.0
numpy==1.16.2
pandas==0.24.2
pkginfo==1.4.2
pluggy==0.8.0
psycopg2==2.7.7
py==1.7.0
pyrsistent==0.14.11
pytest==4.0.0
pytest-django==3.4.4
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.8.0
six==1.12.0
sqlparse==0.2.4
twine==1.11.0
urllib3==1.24.1
yann==0.0.22
